+++ src/heatmap/index.ts
@@ -18,48 +18,6 @@ import {IMatrix, IHeatMapUrlOptions} from 'phovea_core/src/matrix';
 import {IVector} from 'phovea_core/src/vector';
 import {toSelectOperation, defaultSelectionType} from 'phovea_core/src/idtype';
 
-
-export function defaultColor(value: IValueTypeDesc): string[] {
-  if (value.type === VALUE_TYPE_CATEGORICAL) {
-    return (<ICategoricalValueTypeDesc>value).categories.map((c) => typeof c === 'string' ? 'gray' : (<ICategory>c).color || 'gray');
-  }
-  const nv = <INumberValueTypeDesc>value;
-  const r = nv.range;
-  if (r[0] < 0 && r[1] > 0) {
-    //use a symmetric range
-    return ['blue', 'white', 'red'];
-  }
-  return ['white', 'red'];
-}
-export function defaultDomain(value: IValueTypeDesc): (string|number)[] {
-  if (value.type === VALUE_TYPE_CATEGORICAL) {
-    return (<ICategoricalValueTypeDesc>value).categories.map((c) => typeof c === 'string' ? <string>c : (<ICategory>c).name);
-  }
-  const nv = <INumberValueTypeDesc>value;
-  const r = nv.range;
-  if (r[0] < 0 && r[1] > 0) {
-    //use a symmetric range
-    return [Math.min(r[0], -r[1]), 0, Math.max(-r[0], r[1])];
-  }
-  return r;
-}
-
-interface IScale {
-  (x: any): any;
-  domain(): any[];
-  domain(values: any[]): IScale;
-
-  range(): any[];
-  range(values: any[]): IScale;
-}
-
-export function toScale(value): IScale {
-  if (value.type === 'categorical') {
-    return d3.scale.ordinal();
-  }
-  return d3.scale.linear();
-}
-
 enum ESelectOption {
   CELL,
   ROW,
@@ -480,22 +438,6 @@ function createRenderer(d: IMatrix, selectAble: ESelectOption = ESelectOption.CE
   }
 }
 
-export interface ICommonHeatMapOptions extends IVisInstanceOptions {
-  /**
-   * @default 10
-   */
-  initialScale?: number;
-  /**
-   * @default derived from value
-   */
-  color?: string[];
-  /**
-   * @default derived from value
-   */
-  domain?: (number|string)[];
-}
-
-
 export interface IHeatMapOptions extends ICommonHeatMapOptions {
   /**
    * @default null
@@ -636,151 +578,7 @@ export class HeatMap extends AVisInstance implements IVisInstance {
   }
 }
 
-export interface IHeatMap1DOptions extends ICommonHeatMapOptions {
-  /**
-   * width
-   * @default 20
-   */
-  width?: number;
-  /**
-   * scale such that the height matches the argument
-   * @default null
-   */
-  heightTo?: number;
-}
 
-export class HeatMap1D extends AVisInstance implements IVisInstance {
-  private readonly $node: d3.Selection<any>;
-  private readonly colorer: IScale;
-
-  private readonly options : IHeatMap1DOptions = {
-    initialScale: 10,
-    width: 20,
-    heightTo: null,
-    scale: [1,1],
-    rotate: 0
-  };
-
-  constructor(public readonly data: IVector, public parent: Element, options: IHeatMap1DOptions = {}) {
-    super();
-    const value = this.data.valuetype;
-    mixin(this.options, {
-      color: defaultColor(value),
-      domain: defaultDomain(value)
-    }, options);
-    this.options.scale = [1, this.options.initialScale];
-    if (this.options.heightTo) {
-      this.options.scale[1] = this.options.heightTo / this.data.dim[0];
-    }
-    this.colorer = toScale(value).domain(this.options.domain).range(this.options.color);
-    this.$node = this.build(d3.select(parent));
-    this.$node.datum(data);
-    assignVis(this.node, this);
-  }
-
-  get rawSize(): [number, number] {
-    const d = this.data.dim;
-    return [this.options.width, d[0]];
-  }
-
-  get node() {
-    return <Element>this.$node.node();
-  }
-
-  option(name: string, val?: any) {
-    if (arguments.length === 1) {
-      return this.options[name];
-    } else {
-      this.fire('option', name, val, this.options[name]);
-      this.fire('option.' + name, val, this.options[name]);
-      this.options[name] = val;
-      switch (name) {
-        case 'color':
-        case 'domain':
-          this.recolor();
-          break;
-      }
-    }
-  }
-
-  locateImpl(range: Range) {
-    const dims = this.data.dim;
-    const height = dims[0];
-
-    function l(r, max, s) {
-      if (r.isAll || r.isNone) {
-        return [0, max * s];
-      }
-      const ex: any = d3.extent(r.asList());
-      return [ex[0] * s, (ex[1] - ex[0] + 1) * s];
-    }
-
-    const yh = l(range.dim(0), height, this.options.scale[1]);
-    return Promise.resolve(rect(0, yh[0], 20, yh[1]));
-  }
-
-  transform(scale?: [number, number], rotate: number = 0) {
-    const bak = {
-      scale: this.options.scale || [1, 1],
-      rotate: this.options.rotate || 0
-    };
-    if (arguments.length === 0) {
-      return bak;
-    }
-    const dims = this.data.dim;
-    const width = this.options.width, height = dims[0];
-    this.$node.attr({
-      width: width * scale[0],
-      height: height * scale[1]
-    }).style('transform', 'rotate(' + rotate + 'deg)');
-    this.$node.select('g').attr('transform', 'scale(' + scale[0] + ',' + scale[1] + ')');
-    const new_ = {
-      scale: scale,
-      rotate: rotate
-    };
-    this.fire('transform', new_, bak);
-    this.options.scale = scale;
-    this.options.rotate = rotate;
-    return new_;
-  }
-
-  private recolor() {
-    const c = this.colorer;
-    c.domain(this.options.domain).range(this.options.color);
-    this.$node.selectAll('rect').attr('fill', (d) => c(d));
-  }
-
-  private build($parent: d3.Selection<any>) {
-    const dims = this.data.dim;
-    const width = this.options.width, height = dims[0];
-    const $svg = $parent.append('svg').attr({
-      width: width,
-      height: height * this.options.initialScale,
-      'class': 'phovea-heatmap'
-    });
-    const $g = $svg.append('g').attr('transform', 'scale(1,' + this.options.initialScale + ')');
-
-    const c = this.colorer;
-    const data = this.data;
-
-
-    data.data().then((arr) => {
-      const $rows = $g.selectAll('rect').data(arr);
-      const onClick = selectionUtil(data, $g, 'rect');
-      $rows.enter().append('rect').on('click', onClick).attr({
-        width: this.options.width,
-        height: 1
-      }).append('title').text(identity);
-      $rows.attr({
-        fill: (d) => c(d),
-        y: (d, i) => i
-      });
-      $rows.exit().remove();
-    });
-    return $svg;
-  }
-
-}
 
 export function create(data: IMatrix|IVector, parent: Element, options?: IHeatMapOptions|IHeatMap1DOptions): AVisInstance {
   if (data.desc.type === 'matrix') {
